refactor: separate components by categories
