+++ src/app/invoices/layout.tsx
@@ -1,5 +1,5 @@
-import { Sidebar } from "@/components/sidebar";
-import { Topbar } from "@/components/topbar";
+import { Sidebar } from "@/components/layout/sidebar";
+import { Topbar } from "@/components/layout/topbar";
 import { Box } from "@mui/material";
 import type { Metadata } from "next";
